pyproject.toml
[tool.poetry]
name = "aicoinxpert"
version = "0.1.0"
description = ""
authors = ["Andy <andy@anotherbrain.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
bs4 = "^0.0.1"
pyarrow = "^12.0.0"
seaborn = "^0.12.2"
jupyter-dash = "^0.4.2"
nbformat = "^5.9.0"
geopandas = "^0.13.2"
spicy = "^0.16.0"
python-dotenv = "^1.0.0"
minio = "^7.1.15"
pydantic = "^1.10.9"
pytest = "^7.4.0"
sqlalchemy = "^2.0.17"
psycopg2-binary = "^2.9.6"
install = "^1.3.5"
black = "^23.3.0"
isort = "^5.12.0"
pylint = "^2.17.4"
wandb = "^0.15.5"
tqdm = "^4.65.0"
roboflow = "^1.1.2"
darknet = "^0.3"
imageai = "^3.0.3"
utils = "^1.0.1"
picsellia = "^6.9.0"
keyboard = "^0.13.5"
cheese = "^0.1.0"
timer = "^0.2.2"
qgrid = "^1.3.1"
jellyfish = "^1.0.0"
flask = {extras = ["async"], version = "^2.3.3"}
flask-sqlalchemy = "^3.0.5"
passlib = "^1.7.4"
jsonify = "^0.5"
flask-script = "2.0.5"
flask-migrate = "^4.0.4"
flask-login = "^0.6.2"
flask-wtf = "^1.1.1"
flask-admin = "^1.6.1"
pytz = "^2023.3"
faker = "^19.3.1"
screeninfo = "^0.8.1"
lxml = "^4.9.3"
numpy = "^1.25.2"
torchvision = "^0.15.2"
torchaudio = "^2.0.2"
tcl = "^0.2"
tk = "^0.1.0"
torchsummary = "^1.5.1"
torchinfo = "^1.8.0"
scikit-learn = "^1.3.1"
clearml = "1.6.2"
opencv-python = "^4.8.1.78"
nvidia-ml-py = "^12.535.133"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
